# Add .NET Core 2.1 and 3.0 perf improvements