--- a/.gitignore
+++ b/.gitignore
@@ -6,3 +6,4 @@ packages
 .tools
 private.snk
 .idea
+BenchmarkDotNet.Artifacts/