--- a/Crc32.NET.Benchmarks/Crc32.NET.Benchmarks.csproj
+++ b/Crc32.NET.Benchmarks/Crc32.NET.Benchmarks.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFrameworks>net461;netcoreapp2.1;netcoreapp3.1</TargetFrameworks>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Crc32.NET\Crc32.NET.Core.csproj" />
+  </ItemGroup>
+
+</Project>