--- a/Crc32.NET/SafeProxy.cs
+++ b/Crc32.NET/SafeProxy.cs
@@ -1,12 +1,14 @@
 ﻿/* This is .NET safe implementation of Crc32 algorithm.
  * This implementation was investigated as fastest from different variants. It based on Robert Vazan native implementations of Crc32C
  * Also, it is good for x64 and for x86, so, it seems, there is no sense to do 2 different realizations.
- * 
+ *
  * Addition: some speed increase was found with splitting xor to 4 independent blocks. Also, some attempts to optimize unaligned tails was unsuccessfull (JIT limitations?).
- * 
- * 
+ *
+ *
  * Max Vysokikh, 2016-2017
  */
+
+using System;
 
 namespace Force.Crc32
 {
@@ -35,6 +37,7 @@ namespace Force.Crc32
 			}
 		}
 
+#if !NETCORE20 && !NETCORE30
 		public uint Append(uint crc, byte[] input, int offset, int length)
 		{
 			uint crcLocal = uint.MaxValue ^ crc;
@@ -52,9 +55,9 @@ namespace Force.Crc32
 					^ table[(5 * 256) + input[offset + 10]]
 					^ table[(4 * 256) + input[offset + 11]];
 
-				var c = table[(11 * 256) + input[offset + 4]] 
-					^ table[(10 * 256) + input[offset + 5]] 
-					^ table[(9 * 256) + input[offset + 6]] 
+				var c = table[(11 * 256) + input[offset + 4]]
+					^ table[(10 * 256) + input[offset + 5]]
+					^ table[(9 * 256) + input[offset + 6]]
 					^ table[(8 * 256) + input[offset + 7]];
 
 				var d = table[(15 * 256) + ((byte)crcLocal ^ input[offset])]
@@ -72,5 +75,49 @@ namespace Force.Crc32
 
 			return crcLocal ^ uint.MaxValue;
 		}
+#else
+        public uint Append(uint crc, byte[] input, int offset, int length)
+        {
+            return Append(crc, input.AsSpan(offset, length));
+        }
+
+        public uint Append(uint crc, ReadOnlySpan<byte> input)
+		{
+			uint crcLocal = uint.MaxValue ^ crc;
+
+			uint[] table = _table;
+			while (input.Length >= 16)
+			{
+				var a = table[(3 * 256) + input[12]]
+					^ table[(2 * 256) + input[13]]
+					^ table[(1 * 256) + input[14]]
+					^ table[(0 * 256) + input[15]];
+
+				var b = table[(7 * 256) + input[8]]
+					^ table[(6 * 256) + input[9]]
+					^ table[(5 * 256) + input[10]]
+					^ table[(4 * 256) + input[11]];
+
+				var c = table[(11 * 256) + input[4]]
+					^ table[(10 * 256) + input[5]]
+					^ table[(9 * 256) + input[6]]
+					^ table[(8 * 256) + input[7]];
+
+				var d = table[(15 * 256) + ((byte)crcLocal ^ input[0])]
+					^ table[(14 * 256) + ((byte)(crcLocal >> 8) ^ input[1])]
+					^ table[(13 * 256) + ((byte)(crcLocal >> 16) ^ input[2])]
+					^ table[(12 * 256) + ((crcLocal >> 24) ^ input[3])];
+
+				crcLocal = d ^ c ^ b ^ a;
+                input = input.Slice(16);
+            }
+
+            var i = 0;
+			while (i < input.Length)
+				crcLocal = table[(byte)(crcLocal ^ input[i++])] ^ crcLocal >> 8;
+
+			return crcLocal ^ uint.MaxValue;
+		}
+#endif
 	}
 }