@force-net 2020-11-10
Thanks for PR. I'll do my best to find time to merge it and publish new package.
I'm alive, but really has problems with spare time