@neon-sunset 2022-08-22
@brantburnett 
JIT does not do any loop unrolling or auto-vectorization as of today. It can do loop cloning for independent operations but this doesn't seem to be applicable in our case.

As for the mentioned PR, its purpose is a little bit different: .NET 7 introduces fully cross-platform vector operations. What I mean by this is that pre-.NET 7 it was necessary to reference exact intrinsics like `Avx2.DoStuff` or `AdvSimd.DoStuffButArm` which meant a lot of code duplication. However, it is now possible to write an algorithm on top of `Vector<T>`/`VectorXXX<T>` and its extensions once, and it will compile to corresponding efficient codegen that uses vector instructions supported on the target platform (there are limitations - using Vector256 on arm64 will cause it to fallback to scalar code instead of producing unrolled Vector128 operations).

The reason I mention this is that https://github.com/dotnet/runtime/pull/61558 appears to be following the same approach ensuring that we can just use a single method for *word-sized input* instead of `CRC32` intrinsic where available or fallback to fast implementation for specific platform.

However, our use case here is a little bit different so I totally agree with you on the suggestion to upstream improvements for `System.IO.Hashing` instead.